+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * [PR #106]: Fixing system handling of cycle creation when some plugin is not selected by the user
 * [PR #113]: Fixing blog post creation
+* [PR #104]: Sending message for user synchronization with the mobile platform (Issue 98) 
+ - Configure `USER_SYNC_QUEUE` queue
 
 ## [1.3.0] - 17/01/2017